MazamaCoreUtils: Utility Functions for Production R Code

A suite of utility functions providing functionality commonly needed for production level projects such as logging, error handling, cache management and date-time parsing. Functions for date-time parsing and formatting require that time zones be specified explicitly, avoiding a common source of error when working with environmental time series.

The MazamaCoreUtils package was created by MazamaScience to regularize our work building R-based web services.

The main goal of this package is to create an internally standardized set of functions that we can use in various systems that are being run operationally. Areas of functionality supported by this package include:

python style logging

simple error messaging

cache management

API key handling

date parsing and formatting

lat/lon validation and uniqueID creation

source code linting

See also

Author

Maintainer: Jonathan Callahan jonathan.s.callahan@gmail.com

Other contributors:

Eli Grosman eligrosman1@gmail.com [contributor]

Spencer Pease spencerpease618@gmail.com [contributor]

Thomas Bergamaschi tbergama@gmail.com [contributor]