Run package checks

When multiple developers are working on a package, it is crucially important that they check their code changes often. After merging changes from multiple developers it is equally important to check the package thoroughly.

The problem is that frequent checks should be quick or developers won't do them while thorough checks are, by nature, slow.

Our solution is to provide shorthand functions that wrap devtools::check() and pass it a variety of different arguments.

check(pkg = ".")

check_fast(pkg = ".")

check_faster(pkg = ".")

check_fastest(pkg = ".")

check_slow(pkg = ".")

check_slower(pkg = ".")

check_slowest(pkg = ".")

Arguments

pkg
Package location passed to devtools::check().

Value

No return.

Details

The table below describes the args passed to devtools::check():

check_slowest()	| manual = TRUE, run_dont_test = TRUE

| args = c("--run-dontrun", "--use-gct")

check_slower()	| manual = TRUE, run_dont_test = TRUE

| args = c("--run-dontrun")

check_slow()	| manual = TRUE, run_dont_test = TRUE

| args = c()

check()	| manual = FALSE, run_dont_test = FALSE

| args = c()

check_fast()	| manual = FALSE, run_dont_test = FALSE

| build_args = c("--no-build-vignettes")

| args = c("--ignore-vignettes")

check_faster()	| manual = FALSE, run_dont_test = FALSE

| build_args = c("--no-build-vignettes")

| args = c("--ignore-vignettes", "--no-examples")

check_fastest()	| manual = FALSE, run_dont_test = FALSE

| build_args = c("--no-build-vignettes")

| args = c("--ignore-vignettes", "--no-examples", "--no-tests")

See also